pyproject.toml
[project]
name = "localstruct-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "l1m-dot-io>=0.1.6",
    "pydantic>=2.10.6",
]
